perf: move icon detail body to client component for thin static shell
Server component now only generates metadata and JSON-LD. All body
rendering (SidebarShell, IconDetailPage, related icons, version
counterpart logic) moves to IconPageClient which reads from the
already-bundled icons.json. RSC payload shrinks from ~160KB to ~1KB
per icon page, cutting static output from 1.6GB to ~40MB for icon/.
